Move colors/text to separate files + use ts path aliases
