+++ app/ui/theme/colors.ts
@@ -1,0 +1,6 @@
+export const colors = {
+  primary: '#871',
+  secondary: 'burlywood',
+  outline: '#0AA',
+  shadow: '#FFF',
+};
