+++ app/ui/theme/index.ts
@@ -1,0 +1,2 @@
+export * from './colors';
+export * from './text';
