+++ app/ui/theme/text.ts
@@ -1,0 +1,17 @@
+import {TextStyle} from 'react-native';
+
+export const text: {
+  small: TextStyle;
+  medium: TextStyle;
+  large: TextStyle;
+} = {
+  small: {
+    fontSize: 10,
+  },
+  medium: {
+    fontSize: 14,
+  },
+  large: {
+    fontSize: 16,
+  },
+};
